requirements.txt
pillow==10.4.0
psycopg2==2.9.9
python-decouple==3.8
django-cors-headers==4.4.0
asgiref==3.8.1                                                                                                                                                                                                                               
django==5.1.1                                                                                                                                                                                                                                
djangorestframework==3.15.2                                                                                                                                                                                                                  
sqlparse==0.5.1   
